Clifford Wolf: 	Added LOCKING to rsbak3
git-svn-id: http://svn.linbit.com/rsbak3/trunk@19 787cde76-47d9-0310-a09f-da94ad1acf18

+++ rsbak3.sh
@@ -174,6 +174,14 @@ cd "$backupdir"             || { echo "$2:$this: prepgen error #1"; echo; exit 1
 mkdir -p "$2/generation_0"  || { echo "$2:$this: prepgen error #2"; echo; exit 1; }
 cd "$2/generation_0"        || { echo "$2:$this: prepgen error #3"; echo; exit 1; }
 
+if [ -f LOCKFILE ]; then
+	echo "Found lockfile $PWD/LOCKFILE:"
+	cat LOCKFILE; echo; exit 1
+fi
+
+echo "Already running with PID $$ [$0 $*]." > LOCKFILE
+trap "rm -f $PWD/LOCKFILE" EXIT
+
 last=$( ls -d [0-9]*.bak 2> /dev/null | tail -1 )
 
 rm -rf "$this.new"
